pyproject.toml
[project]
name = "gemini-rag-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.64.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
